Refactor CSI to csi/ folder
